+++ testsrc/com/blackrook/physics2d/test/CollisionBody.java
@@ -13,37 +13,37 @@ import com.blackrook.physics2d.Shape2D;
 public class CollisionBody implements CollisionBody2D
 {
 	protected Shape2D shape;
-	public float x;
-	public float y;
-	public float vx;
-	public float vy;
-	public float rotation;
+	public double x;
+	public double y;
+	public double vx;
+	public double vy;
+	public double rotation;
 	
 	public CollisionBody(Shape2D shape)
 	{
 		this.shape = shape;
 	}
 	
-	@Override public int getCollisionGroupMask()		{return -1;}
-	@Override public int getCollisionTargetMask()		{return -1;}
+	@Override public long getCollisionGroupMask()		{return -1;}
+	@Override public long getCollisionTargetMask()		{return -1;}
 	@Override public Shape2D getCollisionShape()		{return shape;}
-	@Override public float getCollisionSquaredRadius()	{return (float)shape.getSquareRadius();}
-	@Override public float getCollisionVelocityX()		{return vx;}
-	@Override public float getCollisionVelocityY()		{return vy;}
-	@Override public float getCollisionRotationZ()		{return rotation;}
+	@Override public double getCollisionSquaredRadius()	{return shape.getSquareRadius();}
+	@Override public double getCollisionVelocityX()		{return vx;}
+	@Override public double getCollisionVelocityY()		{return vy;}
+	@Override public double getCollisionRotationZ()		{return rotation;}
 
-	@Override public float getObjectRadius() 			{return (float)shape.getRadius();}
+	@Override public double getObjectRadius() 			{return shape.getRadius();}
 	@Override public boolean useObjectRadius() 			{return false;}
 	@Override public boolean isExcludedFromCollision()	{return false;}
-	@Override public float getObjectCenterX() 			{return x;}
-	@Override public float getObjectCenterY() 			{return y;}
-	@Override public float getObjectHalfWidth() 		{return shape.getHalfWidth();}
-	@Override public float getObjectHalfHeight() 		{return shape.getHalfHeight();}
-	@Override public float getObjectHalfDepth() 		{return 0;}
-	@Override public float getObjectCenterZ() 			{return 0;}
-	@Override public float getObjectSweepX()			{return -vx;}
-	@Override public float getObjectSweepY()			{return -vy;}
-	@Override public float getObjectSweepZ()			{return 0;}
+	@Override public double getObjectCenterX() 			{return x;}
+	@Override public double getObjectCenterY() 			{return y;}
+	@Override public double getObjectHalfWidth() 		{return shape.getHalfWidth();}
+	@Override public double getObjectHalfHeight() 		{return shape.getHalfHeight();}
+	@Override public double getObjectHalfDepth() 		{return 0;}
+	@Override public double getObjectCenterZ() 			{return 0;}
+	@Override public double getObjectSweepX()			{return -vx;}
+	@Override public double getObjectSweepY()			{return -vy;}
+	@Override public double getObjectSweepZ()			{return 0;}
 	@Override public String toString() 					{return shape.toString() + "@("+x+","+y+")";} 
 
 }
